pyproject.toml
[project]
name = "blogus"
version = "0.2.0"
description = "A comprehensive library for crafting, analyzing, and perfecting AI prompts with advanced features"
authors = [{name = "Dipankar Sarkar", email = "me@dipankar.name"}]
readme = "README.md"
requires-python = ">=3.11"
keywords = ["ai", "prompts", "llm", "analysis", "testing", "templates", "batch-processing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "anthropic>=0.34.2",
    "openai>=1.0.0",
    "groq>=0.11.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "click>=8.1.7",
    "litellm>=1.40.0",
    "pyyaml>=6.0.1",
    "tortoise-orm[asyncpg,asyncmy]>=0.21.0",
    "aerich>=0.7.2",
    "aiosqlite>=0.19.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "jinja2>=3.1.5",
    "aiofiles>=24.1.0",
]
all = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "jinja2>=3.1.5",
    "aiofiles>=24.1.0",
]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pre-commit>=3.5.0",
    "black>=24.8.0",
    "flake8>=7.1.1",
    "mypy>=1.11.2",
]

[project.scripts]
blogus = "blogus.interfaces.cli.main:cli"
blogus-web = "blogus.interfaces.web.main:main"

[project.urls]
Homepage = "https://github.com/terraprompt/blogus"
Repository = "https://github.com/terraprompt/blogus"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["blogus"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
asyncio_mode = "auto"
filterwarnings = [
    "ignore::pydantic.warnings.PydanticDeprecatedSince20:litellm.*",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
strict = true

[tool.aerich]
tortoise_orm = "blogus.infrastructure.database.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
